+++ lib/src/workers/file_system_worker.dart
@@ -6,7 +6,6 @@ enum FileOperation {
   copy('copy'),
   move('move'),
   delete('delete'),
-  exists('exists'),
   mkdir('mkdir'),
   list('list');